@@ -1,70 +1,67 @@
 {
-    "name": "gestion_ecole",
-    "version": "0.1.0",
-    "private": true,
-    "scripts": {
-        "dev": "next dev --turbopack",
-        "build": "next build",
-        "start": "next start",
-        "lint": "next lint"
-    },
-    "dependencies": {
-        "@auth/nextjs": "^0.0.0-380f8d56",
-        "@heroicons/react": "^2.2.0",
-        "@hookform/resolvers": "^5.0.1",
-        "@prisma/client": "^6.5.0",
-        "@radix-ui/react-alert-dialog": "^1.1.14",
-        "@radix-ui/react-avatar": "^1.1.10",
-        "@radix-ui/react-checkbox": "^1.3.2",
-        "@radix-ui/react-dialog": "^1.1.14",
-        "@radix-ui/react-dropdown-menu": "^2.1.15",
-        "@radix-ui/react-label": "^2.1.7",
-        "@radix-ui/react-popover": "^1.1.11",
-        "@radix-ui/react-select": "^2.1.7",
-        "@radix-ui/react-slot": "^1.2.3",
-        "@radix-ui/react-tabs": "^1.1.12",
-        "@radix-ui/react-tooltip": "^1.2.7",
-        "bcrypt": "^5.1.1",
-        "bcryptjs": "^3.0.2",
-        "class-variance-authority": "^0.7.1",
-        "clsx": "^2.1.1",
-        "daisyui": "^5.0.9",
-        "date-fns": "^2.30.0",
-        "jspdf": "^3.0.1",
-        "jspdf-autotable": "^5.0.2",
-        "lucide-react": "^0.488.0",
-        "mysql2": "^3.14.0",
-        "next": "15.2.4",
-        "next-auth": "^4.24.11",
-        "next-themes": "^0.4.6",
-        "puppeteer": "^21.5.0",
-        "react": "^19.0.0",
-        "react-day-picker": "^8.10.1",
-        "react-dom": "^19.0.0",
-        "react-hook-form": "^7.56.4",
-        "react-hot-toast": "^2.5.2",
-        "react-to-print": "^3.1.0",
-        "sonner": "^2.0.5",
-        "tailwind-merge": "^3.3.0",
-        "tailwindcss-animate": "^1.0.7",
-        "zod": "^3.24.2"
-    },
-    "devDependencies": {
-        "@eslint/eslintrc": "^3",
-        "@tailwindcss/forms": "^0.5.10",
-        "@types/node": "^20",
-        "@types/react": "^19",
-        "@types/react-dom": "^19",
-        "autoprefixer": "^10.4.16",
-        "eslint": "^9",
-        "eslint-config-next": "15.2.4",
-        "postcss": "^8.4.32",
-        "postcss-preset-env": "^9.3.0",
-        "tailwindcss": "^3.3.6",
-        "ts-node": "^10.9.2",
-        "typescript": "^5"
-    },
-    "prisma": {
-        "seed": "ts-node prisma/seed.ts"
-    }
+  "name": "tailadmin-vue-pro-2.0.1",
+  "version": "2.0.1",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build",
+    "lint": "eslint . --fix",
+    "format": "prettier --write src/"
+  },
+  "dependencies": {
+    "@fullcalendar/core": "^6.1.15",
+    "@fullcalendar/daygrid": "^6.1.15",
+    "@fullcalendar/interaction": "^6.1.15",
+    "@fullcalendar/list": "^6.1.15",
+    "@fullcalendar/timegrid": "^6.1.15",
+    "@fullcalendar/vue3": "^6.1.15",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/typography": "^0.5.16",
+    "@vueuse/core": "^13.3.0",
+    "apexcharts": "^4.4.0",
+    "axios": "^1.10.0",
+    "dropzone": "^6.0.0-beta.2",
+    "file-saver": "^2.0.5",
+    "flatpickr": "^4.6.13",
+    "jsvectormap": "^1.6.0",
+    "lucide-vue-next": "^0.474.0",
+    "pinia": "^3.0.3",
+    "swiper": "^11.2.1",
+    "vue": "^3.5.13",
+    "vue-flatpickr-component": "^11.0.5",
+    "vue-kanban": "^1.8.0",
+    "vue-router": "^4.5.0",
+    "vue-toastification": "^2.0.0-rc.5",
+    "vue3-apexcharts": "^1.8.0",
+    "vuedraggable": "^4.1.0",
+    "vuevectormap": "^2.0.1",
+    "xlsx": "^0.18.5"
+  },
+  "devDependencies": {
+    "@tailwindcss/postcss": "^4.0.0",
+    "@tsconfig/node22": "^22.0.0",
+    "@types/file-saver": "^2.0.7",
+    "@types/node": "^22.10.7",
+    "@vitejs/plugin-vue": "^5.2.1",
+    "@vitejs/plugin-vue-jsx": "^4.1.1",
+    "@vue/eslint-config-prettier": "^10.1.0",
+    "@vue/eslint-config-typescript": "^14.3.0",
+    "@vue/tsconfig": "^0.7.0",
+    "eslint": "^9.18.0",
+    "eslint-plugin-vue": "^9.32.0",
+    "jiti": "^2.4.2",
+    "npm-run-all2": "^7.0.2",
+    "postcss": "^8.5.1",
+    "prettier": "^3.4.2",
+    "sass-embedded": "^1.83.4",
+    "tailwindcss": "^4.0.0",
+    "typescript": "~5.7.3",
+    "vite": "^6.0.11",
+    "vite-plugin-vue-devtools": "^7.7.0",
+    "vue-tsc": "^2.2.0"
+  }
 }
