@@ -1,9 +1,5 @@
-﻿module.exports = {
+export default {
   plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-    'postcss-preset-env': {
-      features: { 'nesting-rules': false }
-    }
-  }
-} 
+    '@tailwindcss/postcss': {},
+  },
+}
